- updated to latest release of libmseed 2.6
WARNING: it does not pass our test suite so don't use this for a productive environment!

Tried for the last couple hours to figure whats going on - its somehow related to the changes in 2011.006: 2.6rc1 (I tried this RC) - especially time related stuff. Additionally tests comparing mseed.readMSTraces with mseed.readMSTracesViaRecords show that the latter fails. My guess its all time related - eventually it's an Windows issue as I didn't test it on a Linux box yet. So if you have time please give it a try.