Demo: Reading of msrecord from Python object (similar to StringIO)
It is based on the libmseed function msr_parse. This can be usefull when
directly transferring or receiving MiniSEED records from the web a database
etc. The core code of this test is extracted from libmseed.py line 266. If
implementing this functinality, probably the best way is therefore to use the
readMSTracesViaRecords function.