fixing failing gse2 tests, the problem was the following:
assertAlmostEquals(first, second), calculates
round(abs(second-first), precision) == 0
In here second is a float, first a UTCDateTime.
On my linux box, the calculation of
float - UTCDateTime is not defined for float,
and it could not be passed to __rsub__ of UTCDateTime,
as this is not defined.

I do not think, float - UTCDateTime makes any sence,
therefore I refactored the test by comparing timestamps
directly.