Speeding up evalrespUsingDifferentLine Separators:
For testing separate line separators the parsing is the important point,
not the number of frequency points to generate. Therefor, I decreased
the nfft points to 512, which has a huge impact on the test performance.

Profiling evalresp with kcachegrind shows that the fir_asym_trans function
is the performance bottleneck with 262000 calls, which scale with the
number of fft points used