Arch specific name for obspy.taup binary lib.
The obspy.taup library file created with the help of f2py now also has a
architecture specific filename. This unfortunately introduced a small
hack in the setup.py.
Also the architecture specific library names now use underscores instead
of dashes. This was done to avoid an issue with the f2py interface
files.