Workaround for awkward MinGW bug
This appears to be a bug within MinGW and I hope this workaround is
stable enough. If it is, we should submit a patch for libmseed but it
still needs some more testing.
Simply using _fstat() returns -1 and does not fill the `errno` variable
so it essentially just quietly fails without raising at compile or link
time. I found some more references to issues with fstat() in MinGW on
the internet so the problem is more wide-spread.

I am using gcc 4.6.2 from MinGW 3.2.0 on Windows XP.