Added option to draw a scale on the dayplot.
If the kwarg `data_unit` is given, a simple scale will be drawn to the
bottom right corner of the graph. There are some issues with certain
interval lengths due to overlapping y-ticks but I cannot think of an
easy way to circumvent them. I will try to come up with a solution.