test_ascii.py now fully works.
Replaced the dictionary used during reading with an ordered dictionary
so the order of the traces during reading becomes predictable. For some
reason this was never an issue with Python 2 but with Python 3 is
occasionally changes the order of the keys with normal dictionaries.
Also closed some previously unclosed files.