Added optional header_byteorder flag to MiniSEED.
If given, the byteorder will be enforced for reading MiniSEED files.
Useful in some rare cases where the automatic detection fails.

The `verbose` kwarg is now also passed down to libmseed. A value of `0`
is equal to the default verbosity, a value of `2` is very verbose and a
value of `-1` suppressed all warnings (errors are still shown). `-1` is
only recommended for testing purposes as all warnings show a serious
problem with a file.

The negative verbosity is achieved with a custom log function.