Option to ignore file size check in reading sac files
Checking for consistency between actual and theoretical
file size (from `npts`) can be disabled both in
`SacIO.ReadSacFile()` and in `sac.core.readSAC()`.
This is obtained by passing the parameter: `fsize=False`.

`sac.core.isSAC()` has been rewritten, in order not to
rely on `npts` to check file format, but rather checking
`nvhdr` and `delta`, as in `SacIO.IsSACfile()`.