Using proper malloc() while writing MSEED data.
The datasamples pointer in writeMSEED() is now filled with a proper
malloc call and not the old and hackish resizing approach. Now also runs
on Python 2.7.4. Fixes #551.

Also changed the MSTraceGroup to a simple MSTrace as it is only used one
trace at a time.

Requires recompilation of the shared obspy.mseed library.