Making setup.py work with OSX compilers.
The clang compiler shipping with newer OSX versions does not honor the
LIBRARY_PATH variable and thus cannot find the gfortran libraries. This
adds a workaround. Works with OSX 10.7 and the latest official XCode
clang version. Also tested with a local gcc installation. Works with
Python 2.6.8 and 2.7.5. Tested with NumPy 1.6.2 and 1.7.1 for each. All
other dependencies where the latest versions from pypi.