Using Python callbacks for logging within libmseed.
ms_log() will now properly raise warnings and errors. An issue is that
now everything with a level of >= 2 will be an exceptions and everything
below a warning. There is no lower info-like level. Thus it will either
be an exception or a warning. I think this is acceptable for almost all
use cases.