skipping test_CatchOutput for python3
see #743

CatchOutput does not work as expected on Py3k. E.g. it does not catch
shared C library prints on Py3k. It also catches the outputs in an OS
dependent manner on Py3k.

This context manager is used for tests only. Therefore, it is save to skip
the test for Py3k.

If we cannot get a working CatchOutput on Py3k, we could also just replace
it with a dummy, as it does not have any influence besides a little bit
cluttered test log.