obspy-runtests: check on module basis if tests can be run
this commit adds a mechanism to check in obspy-runtests on a module
basis if all tests of the corresponding module can be executed or if all
tests of the corresponding module should be skipped.

this also implements this check for obspy.seishub, such that all tests
will immediately be skipped (even doctests) if the server is not
reachable.  until now, an unreachable server resulted in long timeout
wait time due to doctest executing all tests even after first failures
(this behavior can only be altered with doctest.FAIL_FAST starting with
Python 3.4, unfortunately)