doc: Remove local copy of plot_directive.
We already depend on matplotlib.sphinxext.only_directives, which is
needed by plot_directive, but not used by us anywhere. So why not depend
on the thing we actually use, and drop an extra file.

I am aware this breaks the gallery, for the moment. This will be fixed
soon-ish.