Fix writeTSPAIR with NumPy 1.4.
It seems np.savetxt is not too smart about writing to bytestream files.
And it appears that converting the UTCDateTime to a string creates a
Unicode string, which savetxt does not encode. Anyway, dropping the
np.savetxt call seems like it would be faster, since we don't need to
iterate through the data a couple extra times.