Untested support for CSS files in ASCII.
These data encodings are defined on page 33 of the CSS reference schema,
but it's not clear if there are any separators between each value. I
also do not have any samples files, so this bit remains untested. Maybe
someone will provide an example some time.

The reference schema can be found in the Pisces repository:
https://github.com/jkmacc-LANL/pisces/tree/dev/docs/data