seedlink: Ensure libmseed records are freed during parsing.
I think having the MSRecord allocated by ctypes in one function, and
then got out of scope might cause the GC to drop it too early, but I'm
not 100% sure how ctypes holds on to the references. In any case, even
if it's GC'd at the right time, there are internal blockette references
that Python can't see and would be leaked.

Fixes #918.