reducing appveyor build matrix to 64bit py27
appveyor seems to be really slow. It take days to finish
building if there are just a few commits.

Also, a new job will only start if the old one completed
completely. Especially when there are more commits, we
are blocking all new build through the old ones.

Also we are testing all these version with travis. From my
point of view, testing the different versions with travis
and the most used version on windows is better, than waiting
days (yes it is days) for the builds to finish --- to see
if windows builds are correct.