Update core.py
I've finished the IsKelunjiClassic() function. 

Now I need to work out how to upload the test files and put them into a directory under the core.py file.

Can someone point me in the right direction pls :-)

I've worked out that I need to download the GitHup editor for Windows; so that's the next thing I'll be doing.

Cheers for now - and thanks for all the fish.