Better working import reroutings.
I found some edge cases related to a custom waveform plugin. The
import rerouting should now be more stable.

Another issue is that we cannot have relative imports that are not
valid if executed from the old location of the package.

Lets hope this fixes most issues with the import rerouting. These
are all just workarounds as I don't think there is a proper solution
to these problems.