Bugfix for libmseed.
The microsecond offset in Blockette 1001 was written wrong as a
consequence of some rounding happening when writing the time in the
fixed header. This is now anticipated and corrected for.

The SEED standards prefers the offset in Blockette 1001 to be between
-50 and +49 microseconds. The current libmseed version will always write
between 0 and +100 microseconds. This is not wrong and explicitly
allowed by the SEED standard.

Changing this without modifying the start time of the MSRecord struct at
run-time is a bit more challenging and likely not worth the effort.