extend functionality of deprecated_keywords decorator
improves handling multiple deprecated keywords being mapped to the same
new keyword. prior to this commit, behavior was presumably erratic, with
values for either of the old deprecated keywords overwriting themselves.
now the decorator raises an exception if multiple old keywords are
mapped to a single new keyword *and* the user actually specified values
for them.