Revert debug_headers in doctest, change SNULL.
The sac/__init__.py doctest shows a null tr.stats.sac.dist.  This will
only be retained from the file if debug_headers=True.  When I did this,
however, null string values were being passed to the tr.stats values.  I
think it was because null values like '-12345  ' were being compared to
b'-12345 ', which won't be caught.  I changed header.SNULL to '-12345
', so that null-catching will work (hopefully), and not be passed on.