Add get_..._geo to taup
It can be useful to know the locaton (lattitude and
longitude) of points on a raypath, e.g. to plot
CMB bounce points for PcP. This adds two new methods
to TauPyModel which take the location of the source
and recever, and decorate the path with locations.

In order to allow this we add three new methods to
the TauPyModel:

* get_travel_times_geo: this is simply a wrapper
  around get_travel_times, but it takes the
  event location and station position (latitude
  and longitude) instead of the source-station
  distance in degrees.

* get_pierce_points_geo: this also takes the
  event location and station position, but also
  adds the latitude and longitude to all
  pierce_points in the returnd Arrivals object.
  These can be used to plot peice points on a
  map with the basemap plugin to matplotlib,
  for example.

* get_get_ray_paths_geo: this works like
  get_pierce_points_geo, but decorates the ray
  path with location information. This can be
  used for simple tomographic corrections,
  for example.

Three simple tests are added. These are derived
from three existing tests for the methods that
take an angular distance between the source and
recever. We need to add tests for the locations
along paths and at peice points too.

A new attribute, flatteningOfEarth, has been added
to the TauModel class, and this is respected by
the new methods. This is always set to 0.0
(which makes the calculations operate on a sphere)
but opens up the possibility of considering an
elliptical Earth (or Mars) model. However, the
effect of changing the value will only result in
a change in the an angular distance between the
source and recever (for fixed latitude and longitude).
This results in a small change in trave time compared
to the effect of distorting the layers and raypath,
so there is not much point in using this unless
we want to implement the ellipticiy correction
(e.g. Dziewonski and Gilbert, 1975).