add helper class for matplotlib backend switching
can be used as context manager (for a temporary backend switch) and also
provides a staticmethod that can be used for a backend switch without
switching back. for both use cases `sloppy` kwarg can be used to control
how hard we try to switch and whether a warning will be emitted if the
switch didn't succeed.