Make geographiclib a soft dependency
If geographiclib is not installed fall back
on obspy/geodetics (when we can). In order
to do this I've added optional arguments to
calc_vincenty_inverse and gps2dist_azimuth in
order to allow other ellipsoids (or a sphere) to
be used. There is probably scope for refactoring
and optimisation for the spherical case.

If geographiclib is not installed we do not decorate
the arrival paths or pierce points. New functionality
in obspy/geodetics could allow this to be fixed.

I've also added myself to CONTRIBUTORS.txt and
fixed all PEP8 errors.