Tests with and without geographiclib
This duplicates all three of the taup_geo test
cases and (if geographiclib is installed) tests
the code path using geographiclib and the code
path using the geodetics fallback. Implementing
this involves jumping through some hoops (maybe
a helper function in geodetics somewhere would
be useful) but it seems to work.

Currently the fallback functions fail, but this
is better than having a bug that only shows up
if geographiclib is not installed.

This also turns the ImportError in _pierce_points_geo
and _path_geo into a warning to allow testing of
these functions. Finalising the behaviour is left
for later.