Avoid multiple import of HAS_GEOGRAPHICLIB in test_tau
The reason we needed three imports of HAS_GEOGRAPHICLIB
in test_tau.py is that:

     from geodetics import HAS_GEOGRAPHICLIB
     if HAS_GEOGRAPHICLIB:
         ...

is not the same as:

     import geodetics
     if geodetics.HAS_GEOGRAPHICLIB
         ...

do different things. Using the second form means that
we can update HAS_GEOGRAPHICLIB in geodetics and have
the value change when accessed elsewhere. This patch
changes to the second form, saving multiple importing,
setting and resetting of HAS_GEOGRAPHICLIB.

geodetics could reasonably provide a context manager
(so you could test "with geodetics.nogeographiclib():",
or something) - but this seems like overkill for tests
in one file. I think this approach is probably the cleanest
we will find without adding such a feature.