fix DC beachball patch for "alpha" option != 1
until now, DC beachballs used to get plotted by a color-filled circle overlain
with a filled pressure area patch. this is problematic when opaqueness is
reduced because the pressure area becomes a mix of both colors.

this change now uses the already existing "plot_dc()" algorithm to set up two
patches for both tension and pressure and no full circle in background anymore.

unfortunately a similar trick does not seem to work for "plot_mt()", which
seems to always select the easier to plot patches (whether pressure or tension)
and only switches colors if necessary.

also add a warning for the not yet fixed non-DC part of the bug.