No more warnings when reading origins with arrivals mulitple times.
When an existing resource identifier is created again, ObsPy checks the
referred to objects for equality. If that fails it raises a warning as
that resource identifier would then point to two different objects.
Reading the same file twice does not trigger the warning as the objects
should of course be identical.

For origins with arrivals this equality check happened before the
arrivals where added the new origin - the old has arrivals - thus they
did not compare equal. Now the arrivals are added before the equality
test happens.