Add calc_dist_azi to taup_geo
The documentation of taup_geo.calc_dist claimed to
return the distance and azimuth between source and
receiver but only returned the distance. Add a new
function (calc_dist_azi) that returns the (shortest)
distance, azimuth from source to receiver and (back)
azimuth from receiver back to source. Keep calc_dist
as a thin wrapper around this to maintain the current
API.