reftek: paranoid data unpacking
this adds a check before unpacking the data samples that the packet
array is contiguous in memory.
if it is, we use the fast unpacking, incrementing memory addresses.
if it is not, we use the slightly slower but failsafe approach that
looks up the memory address for each packet separately.