reftek: add channel_number to tr.stats.reftek130
otherwise it's impossible to know which traces belong to which channel
number in the original recording setup after reading.
channel number is in the extended header section of DT packets and was
thus missing so far (stats.reftek130 is based on EH packet info which is
the same for all channels)

also make a deepcopy of the header dictionary to make sure we have
separate instances o all sub-objects in the dictionaries.