deb packaging: fix building man pages
 - also copy py2 .so's to source tree for help2man
 - shift man page generation back

we need to copy so files to source tree before man page generation,
otherwise it will fail

 - ignore warnings about timestamped gzip files

warning shows up for every .gz in our package and that makes it hard to
see really significant warnings/errors.
the gzip files this lintian tag is warning about are not zipped during
package building but are in our tests/data/ directories, so the warning
is not even true (the build is still reproducible, as the timestamps are
the same across builds).

fixes #1272