Change sampling rate of waveform write/read round trip test
Changing the sampling rate so that both sampling rate and sampling
interval can be represented exactly when serializing it as a floating
point number.
Some waveform plugins might store sampling interval instead of sampling
rate and when it can not be represented exactly this leads to an
assertion error when reading the file again and comparing with the
original sampling rate.