fix bug in signal.calibration.rel_calib_stack
In signal.calibration.rel_calib_stack there's a bug in spectrum scaling
when using a non-default window overlap fraction (default is 0.5, i.e.
50% overlap). This bug likely has been in there for ages.

The reason was that computation of how many windows will be computed did
not account for used overlap. Fix is trivial, as the FFT helper routine
returns all individual FFTs, we can just look at the shape of the FFT
result.