core: add decorator to unify file format checks
This should be used on all _is_xxx() file format checks and they should
expect a binary buffer to work on. The decorator will make sure that
already opened file or file-like objects are reset to the initial
position, the decorated function just has to ensure the file-like object
is not closed.

Most logic was reused from the mseed opener as it looked like it might
be the most sophisticated input handling.