Expose keep_sac_header keyword
This exposes a keep_sac_header boolean
flag to the write(..., format='SAC') SAC plugin,
which allows the user to write a "fresh" SAC
header on write, instead of attempting to merge
a previous SAC header (if one exists). The
default value is True, which is current behavior.
This is akin to the previous default SAC plugin
behavior, where the filetype will be ib and the
reference time will be the first sample time (and therefor, the b header will be zero).