Default units "DEF" option for sensor response
Adds a "DEF" option (default units) to Inventory.plot_response()
and Trace.remove_response() (plus other helper functions).

Useful if the units for a particular type of sensor (e.g., a
pressure sensor) cannot be converted to displacement, velocity
or acceleration.

This option is available in the underlying library "evalresp"
(see "-u def" in evalresp manpage). This commit simply exposes it via
the ObsPy API.