Return empty string when loc & cha not found
The current behaviour is to set loc and cha as None in
station_cache if Seedlink doesn't include these subtags in its reply.

This causes get_info to break, as station_cache is then searched
through using fname, with loc and cha provided as the name
argument to fname. A value of None for name will raise, as fname
calls os.path.normpath, which expects a string.

If we set loc and cha as empty strings, the function can proceed.
Any missing location/channel codes will be returned as ''. This does
mean it's up to the user to understand that an empty string means
Seedlink didn't reply with the code, and that the stream may not be
available.