FDSN: Don't reject valid hostnames/TLDs
PR #1162 introduced a validation regex to the FDSN client which rejects
TLDs of length greater than 6. According to RFC 1034, TLDs up to length
63 are valid.

The regex also allowed only alphabetic characters in TLDs, which
excludes the usage of Internationalized Domain Names (which start with
xn-- and may include digits).

Neither of these concerns are theoretical - see the list of registered
TLDs here, which include many long examples and many internationalized
examples:

http://data.iana.org/TLD/tlds-alpha-by-domain.txt

The example that prompted me to submit this patch is a private one - we
use the private DNS zone ".private" in our systems, which FDSN was
rejecting for being too long.

Finally, the regex for unqualified hostnames was not allowing hyphens.

This PR adds some test cases to illustrate the issues above, and alters
the URL regex to validate them. I also made a small change to the
validation test to give better feedback on failures.