Added datetime separator support for SEG2
Added a regular expression for separating the date and time strings. Before these were hardcoded with separators, but across the world different standard separators for date and time are being used. I changed it with a regex that will split the time string on any non-numeric character, which will split on any default separator as well as on `hms`. For the date I used a regex that will split on the worldwide accepted separators: oblique stroke (slash), full stop, dot or point (period), hyphen (dash), and space.