Simplify library naming
In Python 3, the extension suffix includes the architecture and the
Python version, so we don't really need to do that manually ourselves.

There's also not really any need for a `lib` prefix, since they're all
in a `lib` directory, and there's no need to 'cleanse' the name since
the caller is private and we know all our extensions are nameed cleanly.