Fix statements in macros
Doing `#define foo(x) (y=x);` will break code like:
```
if (...)
    foo(0);
else
    foo(1);
```
because semicolon in the macro will be the end of the statement in the
`if`, and the semicolon outside the macro will be an empty statement,
meaning the `else` is no longer attached to the `if`.

Similarly, placing multiple statements into a macro will cause braceless
`if` to break as only the first statement will be within the condition.
This occurs a few times with `swapflag` in `parseutils.c`.