response list: do not raise but warn on extrapolation
response list stages can be confined to quite narrow frequency bands,
e.g. when coming from industry standard calibration protocols. in real
world cases response calculation for deconvolution on time series will
for technical reasons likely have to resort on extrapolation outside of
the frequency range defined by the response list stage. while this is
problematic it is easy to counter by appropriate `pre_filt` in response
removal, which essentially zeros out unwanted parts of the signal
spectrum anyway.

currently it is not possible to do this as we forcefully raise an
exception. this commit changes that and converts it into a warning
message that gives back control to the user.