fix a bug in recalculate sensitivity for Response
During the high level method call for recalculation of sensitivity,
which accepts specifying the frequency to use as normalization frequency
or alternatively tries to determine a sane default, a combination of
other low+high level routines get called.
The problem here seems to be that while the above frequency gets
specified as the target frequency, internally the sensitivity frequency
is still left at its original and potentially inappropriate value, and
this seems to get used in the fairly complex method
`get_evalresp_response_for_frequencies`, leading to wrong results in the
first run, which sets the normalization frequency at the very end,
resulting in consecutive runs producing the correct result.
It seems setting the normalization frequency at the very start fixes the
issue.

NRL.get_response() add test for bugfix in recalculate sensitivity