fdsn massdownloader: make channel_priority robust to advertized but missing data
currently (at least when only "weak" availability data is available),
any channels that are in principle available according to metadata
overrule all other channels that come later in the channel_priority
listing, e.g. if the the first item in channel_priority successfully
matches a channel, all other channels are ignored, even if the former
selected channel actually yields "No data available at server" while
some other channels actually do have data but are later in the
channel_priority (see #2794)

Currently the only way to fix this is to first try and download *all*
channels' data that match any of the given channel_priority wildards,
and then at the very end it is evaluated if some higher priority data
was downloaded and lower priority data get deleted again.

This certainly is not ideal, since it might blow up the amount of data
downloaded and subsequently discarded, but it is likely the lesser evil
compared to losing whole stations from the final download result