response: do not report unknown units as type "displacement" to evalresp
this leads to e.g. input units of radian/second on a rotational sensor
with using the unchanged default of `output="VEL"` to have a
differentiation added on top of the response stated in metadata, which
makes no sense.
if unkown units come into play, just report to evalresp as "undefined"
which seems to consistently just use the response from metadata as is,
essentially ignoring "output" parameter