pytest: make it possible to set the image filetype in image tests
This is done with a new marker "image_path_suffix" that can be used on
test functions. Also be strict with markers so that pytest will raise an
error on unknown markers (e.g. typos in markers) and any new markers
used have to be declared in a central spot (in pytest.ini)