PPSD: improve how the time information for processed data is stored
Previously we were saving the start time of a given cut out slice of the
whole trace as the time that was saved for the processing info. In some
cases it seems this can lead to consecutive valid slices not getting
processed because they seemingly overlap with already processed data.
Instead saving the "canonical" start time that was used to cut out the
data, coming from stepping through the dataset with the fixed "step"
width seems like a safer/better option