Faster read of large miniseed files.
This PR can cut down processing times by anorder of magnitude if start
and end time are used (e.g. for SDS reading with time window).

It uses memory mapped data representation and interpolation search
to find the place which actually must be read instead of fully reading
the file.